requirements.txt
fastapi==0.97.0
pydantic==1.10.9
mysql-connector-python==8.0.33
uvicorn==0.22.0